SplashActivity의 로직을 개발하기위한 주석 추가
